Fix -Wformat warning on Windows

+++ src/api-windows.c
@@ -920,7 +920,7 @@ SEXP psll_interrupt(SEXP p, SEXP ctrlc, SEXP interrupt_path) {
   iret = ps__utf8_to_utf16(cinterrupt_path, &wpath);
   if (iret) goto error;
 
-  iret = snprintf(arguments, sizeof(arguments) - 1, "interrupt.exe %d %s", handle->pid,
+  iret = snprintf(arguments, sizeof(arguments) - 1, "interrupt.exe %d %s", (int) handle->pid,
 		  cctrlc ? "c" : "break");
   if  (iret < 0) goto error;
